# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Padel Concierge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List matches with optional filters
      parameters:
        - name: level
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a new match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/suggestions:
    get:
      operationId: getMatchSuggestions
      tags: [matches]
      summary: Get smart match suggestions for current user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchSuggestions"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Not found
    patch:
      operationId: updateMatch
      tags: [matches]
      summary: Update match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatchBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"

  # BOOKINGS
...
  /stats/player/{id}:
    get:
      operationId: getPlayerStats
      tags: [stats]
      summary: Get stats for a specific player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStats"
  /stats/activity:
    get:
      operationId: getActivityLog
      tags: [stats]
      summary: Get recent activity log
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"

  # ADMIN
...
components:
  schemas:
...
    Match:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        time:
          type: string
        clubName:
          type: string
        format:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/MatchPlayer"
        status:
          type: string
        price:
          type: number
        levelMin:
          type: ["string", "null"]
        levelMax:
          type: ["string", "null"]
        matchType:
          type: string
        balanceScore:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - date
        - time
        - clubName
        - format
        - players
        - status
        - price
        - matchType
        - createdAt
    MatchPlayer:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        level:
          type: string
        confirmed:
          type: boolean
        avatar:
          type: ["string", "null"]
      required:
        - userId
        - name
        - level
        - confirmed
    CreateMatchBody:
      type: object
      properties:
        date:
          type: string
        time:
          type: string
        clubName:
          type: string
        format:
          type: string
        playerIds:
          type: array
          items:
            type: integer
        matchType:
          type: string
      required:
        - date
        - time
        - clubName
        - format
        - playerIds
        - matchType
    UpdateMatchBody:
      type: object
      properties:
        status:
          type: string
        format:
          type: string
        clubName:
          type: string
    MatchSuggestions:
      type: object
      properties:
        best:
          $ref: "#/components/schemas/Match"
        balanced:
          $ref: "#/components/schemas/Match"
        challenging:
          $ref: "#/components/schemas/Match"
        easy:
          $ref: "#/components/schemas/Match"
      required: []
...
    PlayerStats:
      type: object
      properties:
        userId:
          type: integer
        matchesPlayed:
          type: integer
        wins:
          type: integer
        winRate:
          type: number
        levelProgress:
          type: number
        winsToNextLevel:
          type: integer
        levelConfidence:
          type: number
        formatBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/FormatCount"
        winTrend:
          type: array
          items:
            $ref: "#/components/schemas/WinTrendEntry"
        frequentPartners:
          type: array
          items:
            $ref: "#/components/schemas/FrequentPartner"
      required:
        - userId
        - matchesPlayed
        - wins
        - winRate
        - levelProgress
        - winsToNextLevel
        - levelConfidence
        - formatBreakdown
        - winTrend
        - frequentPartners
    FormatCount:
      type: object
      properties:
        format:
          type: string
        count:
          type: integer
      required:
        - format
        - count
    WinTrendEntry:
      type: object
      properties:
        date:
          type: string
        winRate:
          type: number
      required:
        - date
        - winRate
    FrequentPartner:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        matchesTogether:
          type: integer
        avatar:
          type: ["string", "null"]
      required:
        - userId
        - name
        - matchesTogether
    ActivityLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        action:
          type: string
        details:
          type: ["string", "null"]
        detailsParams:
          type: ["object", "null"]
          additionalProperties: true
          description: Language-neutral params for client-side i18n rendering. Key is `activityLog.<action>`.
        createdAt:
          type: string
      required:
        - id
        - userId
        - userName
        - action
        - createdAt
...